requirements.txt
streamlit
requests
python-dotenv
pandas
streamlit-autorefresh
